pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "backgroundremover>=0.3.4",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "trimesh>=4.7.1",
]
